+++ app/src/main/java/com/eszdman/photoncamera/OpenGL/Nodes/Sharpen.java
@@ -24,7 +24,7 @@ public class Sharpen extends Node {
         sharpnessLevel = Math.max(0.8f,sharpnessLevel);
         sharpnessLevel = Math.min(2.0f, sharpnessLevel);
         Log.d("PostNode:"+Name, "sharpnessLevel:" + sharpnessLevel + " iso:" + CameraFragment.mCaptureResult.get(CaptureResult.SENSOR_SENSITIVITY));
-        glProg.servar("size", (float)sharpnessLevel);
+        glProg.servar("size", sharpnessLevel);
         glProg.servar("strength", (float)Interface.i.settings.sharpness);
         glProg.setTexture("InputBuffer",Previous.WorkingTexture);
         super.WorkingTexture = new GLTexture(Previous.WorkingTexture.mSize,Previous.WorkingTexture.mFormat,null);
